release: prepare v1.2.0
