+++ lib/client.js
@@ -16,8 +16,8 @@ window.__ModuleLoader__.load({
     const NS = 'vision-router'
     const zh = {
       nav: '视觉路由（自动识图）',
-      desc: '会话模型负责聊天，视觉后端负责看图；两者分开设置 · 面板 v7',
-      quickStartTitle: '先分清两个模型',
+      desc: '会话模型负责聊天，视觉后端负责看图；两者分开设置 · 面板 v8',
+      quickStartTitle: '聊天与看图分别设置',
       quickStartBody: '① 会话/文字模型：在聊天页右下角选择；发图时选带「+ 自动识图」的模型组。② 视觉模型：在本卡片的「视觉后端链」里选择，只负责看图。',
       quickStartLive: '只想换识图模型，就改下面的「视觉后端链」；不会改变聊天页右下角的会话模型。',
       quickStartGuide: '重新查看模型引导',
@@ -110,6 +110,12 @@ window.__ModuleLoader__.load({
       updateSuccess: '更新命令已完成（目标 v{latest}）。请重启 DSH；重启后新版本才会生效。',
       updateActionFailed: '一键更新失败：{error}',
       updateReleaseNotes: '查看更新说明',
+      updateRegistryFallback: '当前配置的 registry 不可用，已自动改用 npm 官方源完成检查。',
+      updateManualTitle: '手动更新',
+      updateManualSource: '源码仓库 / pnpm DSH：',
+      updateManualNpx: '普通 npm / npx DSH：',
+      updateProject: '项目主页',
+      updateReleases: 'Releases',
       save: '保存',
       saving: '保存中…',
       renderFailed: '设置卡渲染失败：',
@@ -178,8 +184,8 @@ window.__ModuleLoader__.load({
     }
     const en = {
       nav: 'Vision Router (auto image understanding)',
-      desc: 'The session model chats; the vision backend sees images. They are configured separately · panel v7',
-      quickStartTitle: 'Know the two model settings',
+      desc: 'The session model chats; the vision backend sees images. They are configured separately · panel v8',
+      quickStartTitle: 'Chat and vision are configured separately',
       quickStartBody: '① Session/text model: choose it from the lower-right chat selector; for image turns use a group marked “+ Auto Vision”. ② Vision model: choose it in this card under “Vision backend chain”; it only handles image understanding.',
       quickStartLive: 'To change only image understanding, edit “Vision backend chain” below; your lower-right session model is not changed.',
       quickStartGuide: 'Show model guide again',
@@ -272,6 +278,12 @@ window.__ModuleLoader__.load({
       updateSuccess: 'The update command completed (target v{latest}). Restart DSH to load the new version.',
       updateActionFailed: 'One-click update failed: {error}',
       updateReleaseNotes: 'View release notes',
+      updateRegistryFallback: 'The configured registry failed; the check succeeded through the official npm registry.',
+      updateManualTitle: 'Manual update',
+      updateManualSource: 'DSH source checkout / pnpm:',
+      updateManualNpx: 'Normal npm / npx DSH:',
+      updateProject: 'Project',
+      updateReleases: 'Releases',
       save: 'Save',
       saving: 'Saving…',
       renderFailed: 'Settings card failed to render: ',
@@ -1446,6 +1458,14 @@ window.__ModuleLoader__.load({
       const updatePanel = () => {
         const result = updateState.result
         const auto = result && result.autoUpdate
+        const profile = auto && typeof auto.profile === 'string' && auto.profile ? auto.profile : 'web'
+        const projectUrl = 'https://github.com/ysr666/dsh-vision-router'
+        const releasesUrl =
+          result && typeof result.releasesUrl === 'string' && result.releasesUrl
+            ? result.releasesUrl
+            : projectUrl + '/releases/latest'
+        const pnpmCommand = 'pnpm dsh plugin --profile ' + profile + ' update dsh-vision-router'
+        const npxCommand = 'npx @deepseek-ai/dsh plugin --profile ' + profile + ' update dsh-vision-router'
         let status
         let failedUpdate = false
         if (updateState.status === 'running') {
@@ -1476,6 +1496,43 @@ window.__ModuleLoader__.load({
             error: selfUpdateState.result && selfUpdateState.result.error ? selfUpdateState.result.error : 'unknown',
           })
         }
+        const showManualHelp =
+          failedUpdate ||
+          selfUpdateState.status === 'error' ||
+          (result && result.ok === true && result.updateAvailable === true && (!auto || auto.supported !== true))
+        const commandStyle = {
+          display: 'block',
+          whiteSpace: 'pre-wrap',
+          wordBreak: 'break-word',
+          padding: '8px 10px',
+          borderRadius: 8,
+          background: 'var(--dsw-alias-bg-layer-3)',
+          color: 'var(--dsw-alias-label-primary)',
+          fontSize: 12,
+        }
+        const manualHelp = showManualHelp
+          ? h('div', { className: 'vr-catalog-error' },
+              h('div', { className: 'vr-label' }, t('updateManualTitle')),
+              auto && auto.reason === 'source-cli-needs-loader'
+                ? h('div', null,
+                    h('p', { className: 'vr-hint' }, t('updateManualSource')),
+                    h('code', { style: commandStyle }, pnpmCommand),
+                  )
+                : null,
+              h('p', { className: 'vr-hint' }, t('updateManualNpx')),
+              h('code', { style: commandStyle }, npxCommand),
+              h('div', { style: { display: 'flex', gap: 8, flexWrap: 'wrap', marginTop: 8 } },
+                h('button', {
+                  type: 'button', className: 'vr-btn',
+                  onClick: () => window.open(projectUrl, '_blank', 'noopener,noreferrer'),
+                }, t('updateProject')),
+                h('button', {
+                  type: 'button', className: 'vr-btn',
+                  onClick: () => window.open(releasesUrl, '_blank', 'noopener,noreferrer'),
+                }, t('updateReleases')),
+              ),
+            )
+          : null
         return h('div', { className: 'vr-field' },
           h('div', { className: 'vr-field-head' },
             h('span', { className: 'vr-label' }, t('updateTitle')),
@@ -1486,6 +1543,9 @@ window.__ModuleLoader__.load({
             }, updateState.status === 'running' ? t('updateChecking') : t('checkUpdate')),
           ),
           status ? h('p', { className: failedUpdate ? 'vr-failed' : 'vr-hint' }, status) : null,
+          result && result.ok === true && result.registryFallbackFrom
+            ? h('p', { className: 'vr-hint' }, t('updateRegistryFallback'))
+            : null,
           result && result.ok === true && result.updateAvailable === true
             ? h('div', { className: 'vr-catalog-error' },
                 h('p', { className: 'vr-hint' }, auto && auto.supported === true ? t('updateInstallHint') : t('updateAutoUnavailable')),
@@ -1498,18 +1558,21 @@ window.__ModuleLoader__.load({
                       ? t('updateRunning')
                       : t('updateNow', { latest: result.latestVersion }))
                   : null,
-                result.releasesUrl
+                releasesUrl
                   ? h('button', {
                       type: 'button', className: 'vr-btn',
                       disabled: selfUpdateState.status === 'running',
-                      onClick: () => window.open(result.releasesUrl, '_blank', 'noopener,noreferrer'),
+                      onClick: () => window.open(releasesUrl, '_blank', 'noopener,noreferrer'),
                     }, t('updateReleaseNotes'))
                   : null,
                 selfUpdateStatus
                   ? h('p', { className: selfUpdateFailed ? 'vr-failed' : 'vr-hint' }, selfUpdateStatus)
                   : null,
               )
-            : null,
+            : selfUpdateStatus
+              ? h('p', { className: selfUpdateFailed ? 'vr-failed' : 'vr-hint' }, selfUpdateStatus)
+              : null,
+          manualHelp,
         )
       }
