+++ lib/self-update.js
@@ -81,19 +81,20 @@ export function detectDshSelfUpdatePlan({
   }
   const cliEntry = safeRealpath(cliArg)
   if (!cliEntry) return { available: false, reason: 'cli-entry-unresolved' }
+  const profile = profileFromArgv(argv)
+  if (!profile) return { available: false, reason: 'profile-unresolved' }
 
   // Running a raw TS/TSX source entry with plain Node may require a loader
   // owned by the source workspace. Do not guess that loader; fall back to the
-  // manual update instructions instead.
+  // manual update instructions instead. Preserve the profile so the UI can
+  // print the exact pnpm command the user should run.
   const extension = path.extname(cliEntry).toLowerCase()
   if (extension === '.ts' || extension === '.tsx') {
-    return { available: false, reason: 'source-cli-needs-loader' }
+    return { available: false, reason: 'source-cli-needs-loader', profile }
   }
 
   const owner = findDshPackageRoot(cliEntry)
   if (!owner) return { available: false, reason: 'unverified-dsh-cli' }
-  const profile = profileFromArgv(argv)
-  if (!profile) return { available: false, reason: 'profile-unresolved' }
   if (typeof execPath !== 'string' || execPath.trim() === '') {
     return { available: false, reason: 'node-entry-missing' }
   }
