+++ tests/client.test.js
@@ -70,6 +70,12 @@ test('the client bundle still loads and registers with the proven injects', () =
 test('model-selection guide separates session and vision models and targets the vision chain', () => {
   const source = readFileSync(new URL('../lib/client.js', import.meta.url), 'utf8')
   assert.equal(source.includes("onboardingStep1Title: '1 · 会话 / 文字模型'"), true)
+  assert.equal(source.includes("quickStartTitle: '聊天与看图分别设置'"), true)
+  assert.equal(source.includes("quickStartTitle: 'Chat and vision are configured separately'"), true)
+  assert.equal(source.includes("updateProject: '项目主页'"), true)
+  assert.equal(source.includes("updateManualSource: '源码仓库 / pnpm DSH：'"), true)
+  assert.equal(source.includes('pnpm dsh plugin --profile '), true)
+  assert.equal(source.includes('npx @deepseek-ai/dsh plugin --profile '), true)
   assert.equal(source.includes("onboardingStep2Body: '打开「设置 → 插件 → Vision Router」"), true)
   assert.equal(source.includes("onboardingStep1Title: '1 · Session / text model'"), true)
   assert.equal(source.includes('Settings → Plugins → Vision Router'), true)
